requirements.txt
torch>=2.0
nnunetv2>=2.6
numpy
nibabel
SimpleITK
Pillow
